Move subclasses to separate files
